+++ tests/units/OptionParserTest.cpp
@@ -112,7 +112,7 @@ TEST( OptionParserTest, CommandLineOptionWithDataNoAssignment )
     ASSERT_EQ( opts[ 2 ].inCommand, 3 );
     ASSERT_TRUE( OptionParser::isString( opts[ 2 ].inField ) );
     ASSERT_FALSE( OptionParser::hasRange( opts[ 2 ].inField ) );
-    ASSERT_EQ( opts[ 2 ].inString.size( ), 2 );
+    ASSERT_EQ( opts[ 2 ].inString.size( ), static_cast<uint32_t>(2) );
     ASSERT_EQ( opts[ 2 ].inString[ 0 ], "some" );
     ASSERT_EQ( opts[ 2 ].inString[ 1 ], "thing" );
 }
@@ -202,7 +202,7 @@ TEST( OptionParserTest, CommandLineOptionMixedWithData )
         ASSERT_EQ( opt.inCommand, 4 );
         ASSERT_TRUE( OptionParser::isString( opt.inField ) );
         ASSERT_FALSE( OptionParser::hasRange( opt.inField ) );
-        ASSERT_EQ( opt.inString.size( ), 3 );
+        ASSERT_EQ( opt.inString.size( ), static_cast<uint32_t>(3) );
         ASSERT_EQ( opt.inString[ 0 ], "thing" );
         ASSERT_EQ( opt.inString[ 1 ], "and" );
         ASSERT_EQ( opt.inString[ 2 ], "somethingelse" );
@@ -232,7 +232,7 @@ TEST( OptionParserTest, CommandLineOptionLongDataInQuote )
     ASSERT_EQ( opt.inCommand, 2 );
     ASSERT_TRUE( OptionParser::isString( opt.inField ) );
     ASSERT_FALSE( OptionParser::hasRange( opt.inField ) );
-    ASSERT_EQ( opt.inString.size( ), 1 );
+    ASSERT_EQ( opt.inString.size( ), static_cast<uint32_t>(1) );
     ASSERT_EQ( opt.inString[ 0 ], scope );
 }
 
@@ -262,7 +262,7 @@ TEST( OptionParserTest, CommandLineOptionDefaultData )
         ASSERT_TRUE( OptionParser::isFreestyle( opt.inField ) );
         ASSERT_TRUE( OptionParser::isString( opt.inField ) );
         ASSERT_FALSE( OptionParser::hasRange( opt.inField ) );
-        ASSERT_EQ( opt.inString.size( ), 1 );
+        ASSERT_EQ( opt.inString.size( ), static_cast<uint32_t>(1) );
         ASSERT_EQ( opt.inString[ 0 ], "freestyle" );
     } while ( false );
 
@@ -272,7 +272,7 @@ TEST( OptionParserTest, CommandLineOptionDefaultData )
         ASSERT_EQ( opt.inCommand, 2 );
         ASSERT_TRUE( OptionParser::isString( opt.inField ) );
         ASSERT_FALSE( OptionParser::hasRange( opt.inField ) );
-        ASSERT_EQ( opt.inString.size( ), 1 );
+        ASSERT_EQ( opt.inString.size( ), static_cast<uint32_t>(1) );
         ASSERT_EQ( opt.inString[ 0 ], scope );
     } while ( false );
 }
@@ -369,7 +369,7 @@ TEST( OptionParserTest, OptionSimpleWithData )
         ASSERT_EQ( opt.inCommand, 3 );
         ASSERT_TRUE( OptionParser::isString( opt.inField ) );
         ASSERT_FALSE( OptionParser::hasRange( opt.inField ) );
-        ASSERT_EQ( opt.inString.size( ), 1 );
+        ASSERT_EQ( opt.inString.size( ), static_cast<uint32_t>(1) );
         ASSERT_EQ( opt.inString[ 0 ], "something" );
     } while ( false );
 }
@@ -417,7 +417,7 @@ TEST( OptionParserTest, OptionWithDataNoAssignment )
         ASSERT_EQ( opt.inCommand, 3 );
         ASSERT_TRUE( OptionParser::isString( opt.inField ) );
         ASSERT_FALSE( OptionParser::hasRange( opt.inField ) );
-        ASSERT_EQ( opt.inString.size( ), 2 );
+        ASSERT_EQ( opt.inString.size( ), static_cast<uint32_t>(2) );
         ASSERT_EQ( opt.inString[ 0 ], "some" );
         ASSERT_EQ( opt.inString[ 1 ], "thing" );
     } while ( false );
@@ -467,7 +467,7 @@ TEST( OptionParserTest, OptionMixedWithData )
         ASSERT_EQ( opt.inCommand, 3 );
         ASSERT_TRUE( OptionParser::isString( opt.inField ) );
         ASSERT_TRUE( OptionParser::hasRange( opt.inField ) );
-        ASSERT_EQ( opt.inString.size( ), 1 );
+        ASSERT_EQ( opt.inString.size( ), static_cast<uint32_t>(1) );
         ASSERT_EQ( opt.inString[ 0 ], "some" );
     } while ( false );
 
@@ -477,7 +477,7 @@ TEST( OptionParserTest, OptionMixedWithData )
         ASSERT_EQ( opt.inCommand, 4 );
         ASSERT_TRUE( OptionParser::isString( opt.inField ) );
         ASSERT_FALSE( OptionParser::hasRange( opt.inField ) );
-        ASSERT_EQ( opt.inString.size( ), 3 );
+        ASSERT_EQ( opt.inString.size( ), static_cast<uint32_t>(3) );
         ASSERT_EQ( opt.inString[ 0 ], "thing" );
         ASSERT_EQ( opt.inString[ 1 ], "and" );
         ASSERT_EQ( opt.inString[ 2 ], "somethingelse" );
@@ -507,7 +507,7 @@ TEST( OptionParserTest, OptionLongDataInQuote )
     ASSERT_EQ( opt.inCommand, 2 );
     ASSERT_TRUE( OptionParser::isString( opt.inField ) );
     ASSERT_FALSE( OptionParser::hasRange( opt.inField ) );
-    ASSERT_EQ( opt.inString.size( ), 1 );
+    ASSERT_EQ( opt.inString.size( ), static_cast<uint32_t>(1) );
     ASSERT_EQ( opt.inString[ 0 ], scope );
 }
 
@@ -535,7 +535,7 @@ TEST( OptionParserTest, OptionLongDataInQuoteWithError )
     ASSERT_EQ( opt.inCommand, 2 );
     ASSERT_TRUE( OptionParser::isString( opt.inField ) );
     ASSERT_FALSE( OptionParser::hasRange( opt.inField ) );
-    ASSERT_EQ( opt.inString.size( ), 1 );
+    ASSERT_EQ( opt.inString.size( ), static_cast<uint32_t>(1) );
     ASSERT_NE( opt.inString[0], scope );
     ASSERT_TRUE( opt.inString[0].endsWith(error) );
 }
@@ -562,7 +562,7 @@ TEST( OptionParserTest, OptionMultipleParametersAsciiChar )
     ASSERT_EQ( opt.inCommand, 1 );
     ASSERT_TRUE( OptionParser::isString( opt.inField ) );
     ASSERT_FALSE( OptionParser::hasRange( opt.inField ) );
-    ASSERT_EQ( opt.inString.size( ), 2 );
+    ASSERT_EQ( opt.inString.size( ), static_cast<uint32_t>(2) );
     ASSERT_EQ( opt.inString[ 0 ], "./file.txt" );
     ASSERT_EQ( opt.inString[ 1 ], "./some/" );
 }
@@ -605,10 +605,10 @@ TEST( OptionParserTest, OptionFiguresWithData )
         ASSERT_EQ( opt1.inField   , opt2.inField );
         ASSERT_EQ( opt1.inRefSetup, opt2.inRefSetup );
 
-        ASSERT_EQ( opt1.inString.size( ), 1 );
+        ASSERT_EQ( opt1.inString.size( ), static_cast<uint32_t>(1) );
         ASSERT_EQ( opt1.inString[ 0 ], "{1, 1}, { 5, 10 }, { 20, 3 }" );
 
-        ASSERT_EQ( opt2.inString.size( ), 3 );
+        ASSERT_EQ( opt2.inString.size( ), static_cast<uint32_t>(3) );
         ASSERT_EQ( opt2.inString[ 0 ], "{1,1}");
         ASSERT_EQ( opt2.inString[ 1 ], "{5,10}");
         ASSERT_EQ( opt2.inString[ 2 ], "{20,3}");
@@ -623,10 +623,10 @@ TEST( OptionParserTest, OptionFiguresWithData )
         ASSERT_EQ( opt1.inField   , opt2.inField );
         ASSERT_EQ( opt1.inRefSetup, opt2.inRefSetup );
 
-        ASSERT_EQ( opt1.inString.size( ), 1 );
+        ASSERT_EQ( opt1.inString.size( ), static_cast<uint32_t>(1) );
         ASSERT_EQ( opt1.inString[ 0 ], "{1, 10}, 5, 3" );
 
-        ASSERT_EQ( opt2.inString.size( ), 3 );
+        ASSERT_EQ( opt2.inString.size( ), static_cast<uint32_t>(3) );
         ASSERT_EQ( opt2.inString[ 0 ], "{1,10}" );
         ASSERT_EQ( opt2.inString[ 1 ], "5" );
         ASSERT_EQ( opt2.inString[ 2 ], "3" );
